op-node: Fixed integer sizes in the channel header
This mainly modifies the channel_frame, but has some knock on effects
as the width of some of the fields have been reduced. The channel
frame code is also changed more than I expected due to differences
in API of working with fixed int vs uvarints in go. Otherwise the
code reads very similarly with using Reader/Writer APIs.

+++ op-node/rollup/derive/channel_bank.go
@@ -136,7 +136,7 @@ func (ib *ChannelBank) IngestData(data []byte) error {
 		}
 
 		ib.log.Debug("ingesting frame", "channel", f.ID, "frame_number", f.FrameNumber, "length", len(f.Data))
-		if err := currentCh.IngestData(f.FrameNumber, f.IsLast, f.Data); err != nil {
+		if err := currentCh.IngestData(uint64(f.FrameNumber), f.IsLast, f.Data); err != nil {
 			ib.log.Debug("failed to ingest frame into channel", "channel", f.ID, "frame_number", f.FrameNumber, "err", err)
 			if done {
 				return nil
